Use button[type=button] in multiselect
If the `new_button` option is supplied to the `multiselect` form widget and the template contains any inputs, pressing Enter while focused an input would actually activate the new button. The browser sets it as the default submit button for the form since it is before the "Save changes" one on the bottom.

The solution is to have `type="button"` attribute.