Fix sortable.
Previously it was changed to POST all items with automatic positions from 0 to 49.
I have reverted to use GET with from & to by default.

You could use data-post-all to use the POST behavior.

Also the changes to the unminified JS were not committed.

The unminified JS is now committed, converted to tabs and minified with UglifyJS.