Add workaround for images with absolute URL
This is a follow-up after https://github.com/OpenBuildings/jam-tart/pull/22

`URL::site()` should not be used for images.

`URL::site()` will strip everything from the URL except the path string.

In general URLs to images should be configured in `config/jam.php`.
You could set the path prefix there as well as domain.

This is a workaround in `jam-tart` itself to sustain backwards compatibility.